Move hooks into util file and update imports/exports
