+++ src/components/Hamburger/index.stories.tsx
@@ -1,13 +1,12 @@
-import * as React from "react";
-import { FC } from "react";
-import { useSwitch } from "../Switch/useSwitch";
+import React, { FC, useState } from "react";
 
 import Hamburger from "./index";
 export default { title: "Hamburger" };
 
 export const test: FC = () => {
-  const [H] = useSwitch(Hamburger, {
-    size: 25,
-  });
-  return <>{H}</>;
+  const [activated, setActivated] = useState(false);
+
+  return (
+    <Hamburger activated={activated} onClick={() => setActivated((a) => !a)} />
+  );
 };
