+++ src/components/Sidebar/index.tsx
@@ -4,7 +4,7 @@ import { animated, useSpring } from "react-spring";
 
 import useOnClickOutside from "use-onclickoutside";
 import { AnimatedProps } from "../types";
-import { useDisableBodyScroll } from "./useDisableBodyScroll";
+import { useDisableBodyScroll } from "../../utils";
 
 type Props = {
   orientation?: "left" | "right";
